refactor: add getModulePaths to utils index
