+++ packages/repack/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './federated';
 export * from './getDirname';
 export * from './getPublicPath';
 export * from './getResolveOptions';
+export * from './getModulePaths';
